Remove unidades and link chairs directly to clinics
