+++ app/Http/Controllers/Admin/ClinicController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\Clinic;
-use App\Models\Plano;
 use App\Rules\Cnpj;
 use Illuminate\Http\Request;
 
@@ -18,8 +17,7 @@ class ClinicController extends Controller
 
     public function create()
     {
-        $planos = Plano::all();
-        return view('admin.clinics.create', compact('planos'));
+        return view('admin.clinics.create');
     }
 
     public function store(Request $request)
@@ -28,18 +26,43 @@ class ClinicController extends Controller
             'nome' => 'required',
             'cnpj' => ['required', new Cnpj],
             'responsavel' => 'required',
-            'plano_id' => 'required|exists:planos,id',
+            'endereco' => 'required',
+            'cidade' => 'required',
+            'estado' => 'required',
+            'contato' => 'required',
+            'horarios' => 'required|array',
         ]);
 
-        Clinic::create($data);
+        $horarios = $data['horarios'];
+        unset($data['horarios']);
+
+        $clinic = Clinic::create($data);
+
+        foreach ($horarios as $dia => $horario) {
+            if (($horario['abertura'] ?? false) && ($horario['fechamento'] ?? false)) {
+                $clinic->horarios()->create([
+                    'clinic_id' => $clinic->id,
+                    'dia_semana' => $dia,
+                    'hora_inicio' => $horario['abertura'],
+                    'hora_fim' => $horario['fechamento'],
+                ]);
+            }
+        }
 
         return redirect()->route('clinicas.index')->with('success', 'Clínica salva com sucesso.');
     }
 
     public function edit(Clinic $clinic)
     {
-        $planos = Plano::all();
-        return view('admin.clinics.edit', compact('clinic', 'planos'));
+        $horarios = $clinic->horarios
+            ->mapWithKeys(fn($h) => [
+                $h->dia_semana => [
+                    'abertura' => $h->hora_inicio,
+                    'fechamento' => $h->hora_fim,
+                ],
+            ])->toArray();
+
+        return view('admin.clinics.edit', compact('clinic', 'horarios'));
     }
 
     public function update(Request $request, Clinic $clinic)
@@ -48,10 +71,29 @@ class ClinicController extends Controller
             'nome' => 'required',
             'cnpj' => ['required', new Cnpj],
             'responsavel' => 'required',
-            'plano_id' => 'required|exists:planos,id',
+            'endereco' => 'required',
+            'cidade' => 'required',
+            'estado' => 'required',
+            'contato' => 'required',
+            'horarios' => 'required|array',
         ]);
 
+        $horarios = $data['horarios'];
+        unset($data['horarios']);
+
         $clinic->update($data);
+        $clinic->horarios()->delete();
+
+        foreach ($horarios as $dia => $horario) {
+            if (($horario['abertura'] ?? false) && ($horario['fechamento'] ?? false)) {
+                $clinic->horarios()->create([
+                    'clinic_id' => $clinic->id,
+                    'dia_semana' => $dia,
+                    'hora_inicio' => $horario['abertura'],
+                    'hora_fim' => $horario['fechamento'],
+                ]);
+            }
+        }
 
         return redirect()->route('clinicas.index')->with('success', 'Clínica atualizada com sucesso.');
     }
