+++ resources/views/admin/cadeiras/edit.blade.php
@@ -16,11 +16,11 @@
         @csrf
         @method('PUT')
         <div>
-            <label class="mb-2 block text-sm font-medium text-gray-700">Unidade</label>
-            <select name="unidade_id" required class="w-full rounded border-[1.5px] border-stroke bg-gray-2 py-3 px-5 text-sm text-black focus:border-primary focus:outline-none">
+            <label class="mb-2 block text-sm font-medium text-gray-700">Clínica</label>
+            <select name="clinic_id" required class="w-full rounded border-[1.5px] border-stroke bg-gray-2 py-3 px-5 text-sm text-black focus:border-primary focus:outline-none">
                 <option value="">Selecione</option>
-                @foreach ($unidades as $unidade)
-                    <option value="{{ $unidade->id }}" @selected(old('unidade_id', $cadeira->unidade_id) == $unidade->id)>{{ $unidade->nome }}</option>
+                @foreach ($clinics as $clinic)
+                    <option value="{{ $clinic->id }}" @selected(old('clinic_id', $cadeira->clinic_id) == $clinic->id)>{{ $clinic->nome }}</option>
                 @endforeach
             </select>
         </div>
