+++ routes/admin.php
@@ -2,7 +2,6 @@
 
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\Admin\ClinicController;
-use App\Http\Controllers\Admin\UnidadeController;
 use App\Http\Controllers\Admin\CadeiraController;
 use App\Http\Controllers\Admin\ProfileController;
 use App\Http\Controllers\Admin\UserController;
@@ -14,7 +13,6 @@ Route::get('/', function () {
 
 Route::resource('clinicas', ClinicController::class)
     ->parameters(['clinicas' => 'clinic']);
-Route::resource('unidades', UnidadeController::class);
 Route::resource('cadeiras', CadeiraController::class);
 Route::resource('perfis', ProfileController::class)->parameters(['perfis' => 'perfil']);
 Route::resource('usuarios', UserController::class);
